:pencil2:  fix: 더미데이터 구조 수정 및 Map 파일 분리
Map 리랜더링 시 다른 컴포넌트 리랜더링 방지를 위해 Map 파일 분리
